+++ tests/trx_parser_spec.rs
@@ -89,6 +89,7 @@ fn should_parse_transaction_payee(#[case] input: &str, #[case] expected: Option<
 #[case("2023-05-15 txn \"Hello world!\" #a", &[])]
 #[case("2023-05-15 txn \"Hello world!\" ^link-a ^link-b", &["link-a", "link-b"])]
 #[case("2023-05-15 txn \"Hello world!\" ^link-a #tag ^link-b", &["link-a", "link-b"])]
+#[case("2023-05-15 txn \"Hello world!\"^link-a#tag^link-b", &["link-a", "link-b"])]
 fn should_parse_transaction_links(#[case] input: &str, #[case] expected: &[&str]) {
     let DirectiveContent::Transaction(trx) = parse_single_directive(input).content else {
         panic!("was not a transaction");
@@ -135,6 +136,7 @@ fn should_parse_transaction_flag(#[case] input: &str, #[case] expected: Option<c
 #[case("2014-04-23 * \"Flight to Berlin\"\n  Expenses:Flights -1230.27 USD\n  Liabilities:CreditCard", &[])]
 #[case("2014-04-23 * \"Flight to Berlin\" #berlin-trip-2014\n  Expenses:Flights -1230.27 USD\n  Liabilities:CreditCard", &["berlin-trip-2014"])]
 #[case("2014-04-23 * #hello-world #2023_05", &["hello-world", "2023_05"])]
+#[case("2014-04-23 * \"without-space\"#hello-world#2023_05", &["hello-world", "2023_05"])]
 fn should_parse_tags(#[case] input: &str, #[case] expected: &[&str]) {
     let expected: HashSet<_> = expected.iter().copied().collect();
     let trx = parse_single_transaction(input);
@@ -409,8 +411,6 @@ fn should_reject_invalid_input(
         "2023-05-19 *\n  Assets:Cash 1 CHF {1 EUR,}",
         "2023-05-19 *\n  Assets:Cash 1 CHF {, 2023-05-19}",
         "2023-05-19 *\n  Assets:Cash 1 CHF {,}",
-        "2014-04-23 * #hello-world#2023_05",
-        "2014-04-23 *#hello-world #2023_05",
         "pushtag#test",
         "pushtag test",
         "pushtag",
@@ -422,6 +422,7 @@ fn should_reject_invalid_input(
     )]
     input: &str,
 ) {
+    println!("{input}");
     let result = parse::<f64>(input);
     assert!(result.is_err(), "{result:#?}");
 }