+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 * Accept `,` as a thousand separator in amounts
+* Accept `.` in transaction links
 
 
 ## [2.0.2] - 2024-01-17