+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
+## [2.0.4] - 2024-01-19
+
 ### Fixed
 
 * Accept negation of grouped expression (example: `-(2 * 3)`)
@@ -501,7 +504,8 @@ Announce new status of the project in readme: The project is now in "maintenance
  - `Parser` iterator type
 
 
-[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.3...HEAD
+[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.4...HEAD
+[2.0.4]: https://github.com/jcornaz/beancount-parser/compare/v2.0.3...v2.0.4
 [2.0.3]: https://github.com/jcornaz/beancount-parser/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/jcornaz/beancount-parser/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0...v2.0.1